requirements.txt
# absl-py==1.2.0
# aiosignal==1.3.1
# anyio==3.6.2
# astunparse==1.6.3
# async-timeout==4.0.2
# attrs==22.2.0
# bidict==0.22.1
# cachetools==5.2.0
# certifi==2022.9.24
# charset-normalizer==2.1.1
# click==8.1.3
# colorama==0.4.6
# contourpy==1.0.5
# cycler==0.11.0
# editdistance==0.6.0
# fasttext==0.9.2
# filelock==3.8.0
# flatbuffers==22.9.24
# fonttools==4.37.4
# frozenlist==1.3.3
# gast==0.4.0
# google-auth==2.16.3
# google-auth-oauthlib==0.4.6
# google-pasta==0.2.0
# grpcio==1.49.1
# h11==0.14.0
# h5py==3.7.0
# idna==3.4
# importlib-metadata==5.0.0
# joblib==1.2.0
# Keras-Preprocessing==1.1.2
# kiwisolver==1.4.4
# libclang==14.0.6
# linecache2==1.0.0
# Markdown==3.4.1
# MarkupSafe==2.1.1
# 
# multidict==6.0.4
# 
# oauthlib==3.2.1
# opt-einsum==3.3.0
# packaging==21.3
# Pillow==9.2.0
# protobuf==3.19.6
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# pybind11==2.10.4
# pydantic==1.10.7
# pyenchant==3.2.2
# pyparsing==3.0.9
# python-dateutil==2.8.2
# python-engineio==4.4.0
# python-version==0.0.2
# pytz==2022.4
# requests==2.28.1
# requests-file==1.5.1
# requests-oauthlib==1.3.1
# rsa==4.9
# six==1.16.0
# smart-open==6.3.0
# sniffio==1.3.0
# starlette==0.20.4
# tensorboard==2.10.1
# tensorboard-data-server==0.6.1
# tensorboard-plugin-wit==1.8.1
# tensorflow==2.10.0
# tensorflow-estimator==2.10.0
# tensorflow-hub==0.12.0
# tensorflow-io-gcs-filesystem==0.27.0
# termcolor==2.0.1
# threadpoolctl==3.1.0
# tldextract==3.4.0
# traceback2==1.4.0
# typing-extensions==4.4.0
# urllib3==1.26.12
# Werkzeug==2.2.2
# wrapt==1.14.1
# yarl==1.8.2
# zipp==3.9.0

py_vncorenlp
transformers
scikit-learn
py_vncorenlp
transformers[torch]
fastapi==0.85.0
fastapi-socketio==0.0.9
aiohttp==3.8.3
argparse
uvicorn==0.18.3
python-socketio==5.0.4
gensim==4.3.1
pyvi
keras==2.10.0
pandas==1.5.0
scipy==1.9.2
tensorflow==2.10.0
tqdm==4.64.1
matplotlib==3.6.1
numpy==1.23.3